Refactored the mail sending out into its own function
