+++ src/api/forgotPassword.js
@@ -3,6 +3,7 @@ const nodemailer = require("nodemailer");
 const User = require("../models/User");
 const { sendResponse } = require("./../utils/sendResponse");
 const { isGmailEnabled } = require("../utils/getConfigFile");
+const { sendMail } = require("../utils/sendMail");
 router.post("/forgotPassword", async function(req, res) {
   const usingGmail = await isGmailEnabled();
   if (!usingGmail) {
@@ -35,30 +36,14 @@ router.post("/forgotPassword", async function(req, res) {
     date.setDate(date.getDate() + 1);
     user.expiration = date;
     await user.save();
-    let transporter = nodemailer.createTransport({
-      service: "gmail",
-      auth: {
-        type: "OAuth2",
-        user: process.env.INFRA_EMAIL,
-        clientId: process.env.INFRA_CLIENT_ID,
-        clientSecret: process.env.INFRA_CLIENT_SECRET,
-        refreshToken: process.env.INFRA_REFRESH_TOKEN
-      }
-    });
-    transporter
-      .sendMail({
-        from: "hack4impact.infra@gmail.com",
-        to: user.email,
-        subject: "Forgot Password",
-        text: "Enter the following pin on the reset page: " + user.pin
-      })
-      .catch(_ => {
-        sendResponse(
-          res,
-          500,
-          "An internal server error occured and the email could not be sent."
-        );
-      });
+
+    const body = {
+      from: "hack4impact.infra@gmail.com",
+      to: user.email,
+      subject: "Forgot Password",
+      text: "Enter the following pin on the reset page: " + user.pin
+    };
+    await sendMail(body);
     sendResponse(res, 200, "Sent password reset PIN to user if they exist");
   } else {
     sendResponse(res, 200, "Answer to security question doesn't match");
